+++ win32/Lib/win32con.py
@@ -1,4 +1,4 @@
-from win32usercon import *  # For backwards compatibility
+from winusercon import *  # For backwards compatibility
 
 # Generated by h2py from commdlg.h (plus modifications 4jan98)
 PY_0U = 0
